feat(vindex3): pin publish revision + V3-safe publish defaults
Two small, independently-useful publish fixes, grounded and confirmed
in docs/vindex3-registry-publishing-design.md ahead of the registry
schema itself:

1. publish_vindex/publish_vindex_with_opts now return
   PublishResult{url, revision} instead of a bare String. A new
   fetch_repo_head_sha (publish/remote.rs) makes one extra
   repo.info() call immediately after the last per-file upload commit
   lands, reading main's HEAD sha. Hard error if the fetch fails —
   never falls back to a floating "main" reference, matching the
   registry's "revision must be an immutable pin" rule. `larql
   publish`/`larql hf publish` now print the pinned revision. This is
   what makes a mechanically-derived RegistryArtifactRef::revision
   possible at all.

2. `larql publish` no longer crashes on a VINDEX3 container by
   default. publish_cmd::run detects generation up front and, for V3
   only, downgrades --slices/--collections to their `none` equivalent
   unless the caller passed them explicitly, printing why. VINDEX2
   behaviour is unchanged. A single DEFAULT_COLLECTIONS const now
   feeds both the clap default and the runtime check, closing a
   pre-existing literal duplication too.

Alongside: file-splitting cleanup so every touched file stays under
the 800-line cap. publish_cmd.rs -> publish_cmd/{mod,collections,
upload,progress}.rs. publish/{mod.rs,remote.rs} each gained a sibling
*_tests.rs (format/generation.rs+generation_tests.rs pattern).
download/tests.rs split into tests.rs (core hf_hub plumbing) +
tests_v3.rs (VINDEX3-completeness acceptance tests) + test_support.rs
(scaffolding shared by both). Splitting remote.rs surfaced a real,
previously-diluted coverage gap in its transport/JSON-parse error
paths (six .send() failures, three resp.json() failures) — closed
with 9 new tests instead of lowering the coverage-policy.json
baseline: ordinary mockito for the JSON-parse branches, a
127.0.0.1:0 unreachable endpoint (fails instantly, no timeout) for
the transport branches.

Gates: larql-vindex 2742 lib tests + all integration binaries,
larql-cli 29/29 publish_cmd::, larql-server 575 lib, all green;
clippy -D warnings and fmt --check clean on larql-vindex/larql-cli;
scripts/check_coverage_policy.py clean except the same pre-existing,
untouched quant/convert.rs drift (81.54% vs 81.90%, unrelated to this
change). Full workspace build clean.
