+++ crates/larql-vindex/src/format/huggingface/download/tests.rs
@@ -5,6 +5,11 @@
 
 //! Unit tests for the hf_hub-bound functions — pure helpers tested
 //! in `helpers.rs`.
+//!
+//! VINDEX3 generation-aware download-completeness tests live in the
+//! sibling `tests_v3.rs`; both files share the mock/env scaffolding in
+//! `test_support.rs`.
+use super::test_support::{mock_hf_file_resolve, HfTestEnv, NoOpProgress};
 use super::*;
 use serial_test::serial;
 
@@ -39,16 +44,6 @@ fn download_hf_weights_rejects_empty_string() {
     assert!(err.to_string().contains("not an hf://"));
 }
 
-/// Stub `DownloadProgress` for the *_with_progress tests. We only need
-/// the trait to exist so the function type-checks; the stub is never
-/// invoked because we hit the early-return path.
-struct NoOpProgress;
-impl DownloadProgress for NoOpProgress {
-    fn init(&mut self, _size: usize, _filename: &str) {}
-    fn update(&mut self, _size: usize) {}
-    fn finish(&mut self) {}
-}
-
 #[test]
 fn resolve_hf_vindex_with_progress_rejects_non_hf_path() {
     let err =
@@ -69,56 +64,6 @@ fn resolve_hf_model_with_progress_rejects_non_hf_path() {
 // before any network call. The functions then do `Api::new()` which
 // (with HF_ENDPOINT pointing at a non-existent server) fails fast.
 // That path covers the revision-vs-no-revision branches.
-
-/// RAII guard for HF_ENDPOINT + HF_HOME + a tempdir cache.
-/// Restores prior values on drop.
-struct HfTestEnv {
-    prev_endpoint: Option<String>,
-    prev_home: Option<String>,
-    prev_hub: Option<String>,
-    prev_token: Option<String>,
-    // Hold the tempdir so it lives as long as the guard.
-    _tmp: tempfile::TempDir,
-}
-impl HfTestEnv {
-    fn new(endpoint: &str) -> Self {
-        let prev_endpoint = std::env::var("HF_ENDPOINT").ok();
-        let prev_home = std::env::var("HF_HOME").ok();
-        let prev_hub = std::env::var("HUGGINGFACE_HUB_CACHE").ok();
-        let prev_token = std::env::var("HF_TOKEN").ok();
-
-        let tmp = tempfile::tempdir().unwrap();
-        std::env::set_var("HF_ENDPOINT", endpoint);
-        std::env::set_var("HF_HOME", tmp.path());
-        // Clear HUGGINGFACE_HUB_CACHE so HF_HOME wins; clear token
-        // so we don't accidentally hit a real auth header.
-        std::env::remove_var("HUGGINGFACE_HUB_CACHE");
-        std::env::remove_var("HF_TOKEN");
-
-        Self {
-            prev_endpoint,
-            prev_home,
-            prev_hub,
-            prev_token,
-            _tmp: tmp,
-        }
-    }
-}
-impl Drop for HfTestEnv {
-    fn drop(&mut self) {
-        for (k, prev) in [
-            ("HF_ENDPOINT", self.prev_endpoint.take()),
-            ("HF_HOME", self.prev_home.take()),
-            ("HUGGINGFACE_HUB_CACHE", self.prev_hub.take()),
-            ("HF_TOKEN", self.prev_token.take()),
-        ] {
-            match prev {
-                Some(v) => std::env::set_var(k, v),
-                None => std::env::remove_var(k),
-            }
-        }
-    }
-}
 
 #[test]
 #[serial]
@@ -619,38 +564,6 @@ fn cache_prefix_matches_repo_kind() {
     assert_eq!(RepoKind::Model.cache_prefix(), "models--");
 }
 
-/// Build a mock that satisfies hf-hub's metadata() probe. The
-/// requirements are: an ETag (or X-Linked-Etag) header, an
-/// X-Repo-Commit header (commit hash), and a Content-Range header
-/// of the form "bytes 0-0/<size>". hf-hub's metadata path issues a
-/// GET with `Range: bytes=0-0`, then a follow-up GET for the full
-/// body; both must succeed for `repo.get()` to return a path.
-///
-/// `expect` lets the test cap how many times the mock can fire —
-/// hf-hub's download path may retry or make follow-up requests we
-/// don't strictly model.
-fn mock_hf_file_resolve(
-    server: &mut mockito::ServerGuard,
-    path_regex: &str,
-    etag: &str,
-    body: &[u8],
-) -> Vec<mockito::Mock> {
-    let len = body.len();
-    let cr = format!("bytes 0-0/{len}");
-    let meta = server
-        .mock("GET", mockito::Matcher::Regex(path_regex.into()))
-        .with_status(200)
-        .with_header("ETag", &format!("\"{etag}\""))
-        .with_header("X-Repo-Commit", "deadbeefcafebabe")
-        .with_header("Content-Range", &cr)
-        .with_header("Accept-Ranges", "bytes")
-        .with_header("Content-Length", &len.to_string())
-        .with_body(body)
-        .expect_at_least(1)
-        .create();
-    vec![meta]
-}
-
 #[test]
 #[serial]
 fn resolve_hf_vindex_success_via_broad_mocks() {
@@ -816,165 +729,4 @@ fn cached_snapshot_file_with_revision_falls_back_to_pinned_dir() {
         cached_snapshot_file(RepoKind::Dataset, "owner/repo", Some("v3"), "f.bin").unwrap();
     assert_eq!(size, 3);
     assert!(path.ends_with("f.bin"));
-}
-
-// ─── VINDEX3 generation-aware completeness (vindex3-registry 2C) ───────
-//
-// The acceptance test that rung cares about most: not "these expected
-// filenames were requested" (a hardcoded-list assertion the original
-// bug would have passed too, since the list just needed the *wrong*
-// names) but "the pulled result actually opens as a complete VINDEX3
-// container." A real, self-encoded fixture is served over a mocked HF
-// endpoint; every file it holds is discovered dynamically by walking
-// the fixture on disk, never hardcoded here — so this test cannot
-// silently drift the way the fixed VINDEX2 list did the day VINDEX3
-// grew a payload it didn't know about.
-
-/// Recursively collect every file under `root` as `(relative POSIX
-/// path, bytes)`, sorted. Drives both the mock (what exists, what bytes
-/// it holds) and the `repo.info()` siblings list — from the same
-/// source, so the test can't accidentally list files the mock doesn't
-/// actually serve or vice versa.
-fn walk_fixture_files(root: &std::path::Path) -> Vec<(String, Vec<u8>)> {
-    fn walk(dir: &std::path::Path, root: &std::path::Path, out: &mut Vec<(String, Vec<u8>)>) {
-        for entry in std::fs::read_dir(dir).unwrap() {
-            let entry = entry.unwrap();
-            let path = entry.path();
-            if path.is_dir() {
-                walk(&path, root, out);
-            } else {
-                let rel = path
-                    .strip_prefix(root)
-                    .unwrap()
-                    .to_str()
-                    .unwrap()
-                    .replace('\\', "/");
-                out.push((rel, std::fs::read(&path).unwrap()));
-            }
-        }
-    }
-    let mut out = Vec::new();
-    walk(root, root, &mut out);
-    out.sort();
-    out
-}
-
-#[test]
-#[serial]
-fn resolve_hf_vindex_with_progress_downloads_a_complete_v3_container() {
-    // A real, self-encoded VINDEX3 fixture — segments, moe_manifest.json,
-    // the actual container shape — not a hand-written `{"version": 3}`
-    // stub like a filename-list test would settle for.
-    let checkpoint = tempfile::tempdir().unwrap();
-    let container = tempfile::tempdir().unwrap();
-    crate::format::vindex3::fixtures::encode_fixture_container(
-        crate::format::vindex3::fixtures::miniature_glimmer,
-        checkpoint.path(),
-        container.path(),
-        "pull-fixture",
-    );
-    let files = walk_fixture_files(container.path());
-    assert!(
-        files.len() > 1,
-        "fixture must carry more than index.json for this test to mean anything"
-    );
-
-    let mut server = mockito::Server::new();
-    let _g = HfTestEnv::new(&server.url());
-
-    // `repo.info()` drives the V3 branch's enumeration — list exactly
-    // what the fixture wrote, discovered above, never hardcoded here.
-    let siblings: Vec<_> = files
-        .iter()
-        .map(|(rel, _)| serde_json::json!({"rfilename": rel}))
-        .collect();
-    let info_body = serde_json::json!({"sha": "deadbeef", "siblings": siblings}).to_string();
-    let _info = server
-        .mock(
-            "GET",
-            mockito::Matcher::Regex(r"/api/models/owner/repo".into()),
-        )
-        .with_status(200)
-        .with_header("Content-Type", "application/json")
-        .with_body(info_body)
-        .create();
-
-    let mocks: Vec<_> = files
-        .iter()
-        .map(|(rel, bytes)| {
-            let regex = rel.replace('.', r"\.");
-            mock_hf_file_resolve(&mut server, &regex, rel, bytes)
-        })
-        .collect();
-
-    let dir = resolve_hf_vindex_with_progress("hf://owner/repo", |_| NoOpProgress)
-        .expect("a complete VINDEX3 repo must download in full");
-
-    for (mock, (rel, _)) in mocks.iter().zip(&files) {
-        assert!(mock[0].matched(), "{rel} must have been fetched");
-    }
-
-    // The acceptance test itself: open the PULLED result through the
-    // real VINDEX3 loader, not a filename checklist. `encode_fixture_container`
-    // produces a system-graph container, not a routed-MoE one — its
-    // real loader is `inspect_container` (`Vindex3Container::open`
-    // itself refuses non-MoE containers by name, directing callers
-    // here; see its own error text). `verify_payloads: true` also
-    // checks the downloaded segment bytes are readable and
-    // correctly shaped, not just that the manifest parses.
-    crate::format::vindex3::inspect::inspect_container(&dir, true)
-        .expect("a fully-downloaded VINDEX3 container must open and verify cleanly");
-}
-
-#[test]
-#[serial]
-fn resolve_hf_vindex_with_progress_v3_hard_fails_when_a_listed_file_is_missing() {
-    // `repo.info()` lists a segment file the mock never actually
-    // serves (a corrupt/incomplete upload) — must be a hard failure
-    // naming the file, never a silently-incomplete "success".
-    let checkpoint = tempfile::tempdir().unwrap();
-    let container = tempfile::tempdir().unwrap();
-    crate::format::vindex3::fixtures::encode_fixture_container(
-        crate::format::vindex3::fixtures::miniature_glimmer,
-        checkpoint.path(),
-        container.path(),
-        "pull-fixture",
-    );
-    let files = walk_fixture_files(container.path());
-
-    let mut server = mockito::Server::new();
-    let _g = HfTestEnv::new(&server.url());
-
-    // No catch-all mock — as elsewhere in this file, mockito's default
-    // response for an unmatched request (the deliberately-unserved file
-    // below) is sufficient; an explicit `Matcher::Any` mock risks
-    // shadowing the specific per-file mocks registered around it.
-
-    let mut siblings: Vec<_> = files
-        .iter()
-        .map(|(rel, _)| serde_json::json!({"rfilename": rel}))
-        .collect();
-    // Claim a file the mock will never serve.
-    siblings.push(serde_json::json!({"rfilename": "routed/layer_999.lyrw"}));
-    let info_body = serde_json::json!({"sha": "deadbeef", "siblings": siblings}).to_string();
-    let _info = server
-        .mock(
-            "GET",
-            mockito::Matcher::Regex(r"/api/models/owner/repo".into()),
-        )
-        .with_status(200)
-        .with_header("Content-Type", "application/json")
-        .with_body(info_body)
-        .create();
-    for (rel, bytes) in &files {
-        let regex = rel.replace('.', r"\.");
-        mock_hf_file_resolve(&mut server, &regex, rel, bytes);
-    }
-
-    let err = resolve_hf_vindex_with_progress("hf://owner/repo", |_| NoOpProgress)
-        .expect_err("a repo missing a listed file must hard-fail, not silently succeed");
-    assert!(
-        err.to_string().contains("routed/layer_999.lyrw"),
-        "error must name the missing file: {err}"
-    );
 }
