+++ crates/larql-vindex/src/lib.rs
@@ -101,7 +101,7 @@ pub use format::huggingface::{
     is_hf_path, publish_vindex, publish_vindex_with_opts, repo_exists,
     resolve_hf_model_with_progress, resolve_hf_vindex, resolve_hf_vindex_complete,
     resolve_hf_vindex_with_progress, set_repo_visibility, CollectionItem, DownloadProgress,
-    PublishCallbacks, PublishOptions, SilentPublishCallbacks,
+    PublishCallbacks, PublishOptions, PublishResult, SilentPublishCallbacks,
 };
 pub use format::weights::{
     arch_from_vindex_config, load_model_weights, load_model_weights_kquant,
